requirements.txt
###Base:
cffi==1.15.1
pycparser==2.21
idna==3.3
py4j==0.10.9.5
pipdeptree==2.13.2   ## is a command line utility for displaying the installed python packages in form of a dependency tree. 
typing_extensions==4.12.2
six==1.16.0
ipykernel==6.29.5
loguru==0.7.2

###HTTP and web Connections:
certifi==2025.1.31
charset-normalizer==2.0.4
cryptography==37.0.1
oauthlib==3.2.0
msrest==0.6.21
requests-oauthlib==1.3.1
urllib3==1.26.11
requests==2.32.3

####time:
pytz==2022.1
isodate==0.6.1
python-dateutil==2.8.2
schedule==1.2.2

###Basic Data Science:
pandas==2.2.3
scipy==1.9.1
numpy==1.24.4
statsmodels==0.14.4 
pyarrow==17.0.0
scikit-learn==1.6.0
tqdm==4.66.5
iterative-stratification==0.1.9
ydata-profiling==4.16.1

####I/O:
python-dotenv==1.1.1
fastparquet==2024.2.0
pyodbc==5.1.0
pyYAML==6.0.1
openpyxl==3.1.5
XlsxWriter==3.2.0
#Spire.Doc==13.1.0
xlrd==2.0.1
sshtunnel==0.4.0
psycopg2-binary==2.9.10

####plot:
matplotlib==3.7.3
seaborn==0.11.2
kaleido==0.2.1
nbformat==5.10.4
plotly==5.24.1
ipywidgets== 8.1.7

###ML models:
lightgbm==3.3.4
shap==0.44.1
xgboost==2.1.4
#yellowbrick==1.5
#"u8darts[all]"==0.27.2

####ML Tuning:
hyperopt==0.2.7
optuna==3.5.0

####NLP and LLM:
spacy>=3.4.0
transformers>=4.20.0
sentence-transformers>=2.2.0
langchain>=0.0.200
openai>=0.27.0
tiktoken>=0.4.0
setfit>=0.7.0

####RAG and Document Processing:
docling>=1.0.0
pypdfium2>=4.0.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
chromadb>=0.4.0
faiss-cpu>=1.7.0

####Azure and Cloud:
azure-storage-blob>=12.14.0
azure-keyvault-secrets>=4.6.0
azure-identity>=1.12.0
azure-synapse-spark>=0.7.0

####Spark:
pyspark>=3.3.0

####Platform Integration:
google-colab-tools>=0.1.0
kaggle>=1.5.0
databricks-cli>=0.17.0
ipython>=8.0.0

####Statisical Analysis:
##rpy2==3.5.13
##lifelines==0.28.0
##researchpy==0.3.5  ###outdated remove any dependency

#cd C:\ufolder\contains\requirements.txt
#echo yes Y | conda create --name ml_env python=3.9
#conda activate ml_env
#pip install -r  requirements.txt --user
#python -m ipykernel install --user --name ml_env --display-name "ml_env"

###Remove Conda:
##conda remove -n ml_env --all
##rmdir /S C:\Users\rnourzadeh\AppData\Local\miniconda3\envs\ml_env
